server/requirements.txt
requests==2.12.4
flask==0.12
flask-restful==0.3.5
influxdb==4.0.0
python-dateutil==2.6.0
pytz==2016.10

# dev
ipython
pudb
